requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
